refactor(endor): hoist RunInput + classify_run_input to lib module
Move the `RunInput` enum and `classify_run_input` function from
`src/bin/endor.rs` into `src/run_input.rs` so they are reachable
from `cargo test --lib`. The binary keeps the same public-CLI
behavior: `endor run <path>` dispatches the same three paths
based on the same classification rules.

The hoist makes the Phase 4 magic-byte fallback (the new
`PK\x03\x04` detection rule that lets `endor run foo` work when
`foo` is an extension-less ZIP) testable without spawning the
built binary. The follow-up test commit adds eight cases that
exercise the classifier directly.

No behavior change; pure code move plus `pub`/`#[derive]` adds on
the moved items so the bin can import them and tests can assert
equality.
